Replace uintptr & eliminate bounds-checking risks
Upgrade golangci-lint to v2.10.1.

Eliminate unnecessary or bounds-unchecked integer conversions that gosec
reports at risk of overflow from the lexer. Replace use of the internal
`uintptr` type with `uint` and panic if the difference between two
pointers exceeds int64 max (which should never happen).

While at it, add CI testing for Go 1.25 and 1.26.

+++ path/exec/const_test.go
@@ -14,7 +14,7 @@ func TestExecConstNode(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 	path, _ := parser.Parse("$")
-	base := kvBaseObject{addr: uintptr(42), id: -1}
+	base := kvBaseObject{addr: uint(42), id: -1}
 	current := []any{"hi", true}
 
 	for _, tc := range []struct {
